requirements.txt
Flask~=3.1.0
flasgger~=0.9.7.1
Flask-Migrate~=4.0.7
PyMySQL~=1.1.1
bcrypt~=4.2.1
SQLAlchemy~=2.0.36
PyJWT~=2.10.1